+++ Changelog.md
@@ -16,6 +16,7 @@
 - Fix: Jade/TOP/WTHIT show the owner of block mines
 - Fix: Several instances of blocks disregarding the team ownership config setting being disabled when FTB Teams is installed
 - Fix: Security Camera entities are sent to every player instead of only the player mounting the camera
+- Fix: Trees cannot grow on reinforced dirt-like blocks
 - Misc.: Salt keys are now stored as integer arrays instead of strings, reducing file size
 
 --------------------------Changelog for v1.9.11 of SecurityCraft--------------------------